Extract id as separate class
