+++ src/main/java/seedu/address/model/project/ProjectId.java
@@ -1,24 +1,12 @@
 package seedu.address.model.project;
 
-import static java.util.Objects.requireNonNull;
-import static seedu.address.commons.util.AppUtil.checkArgument;
+import seedu.address.model.id.Id;
 
 /**
  * Represents a Project's ID in the address book.
  * Guarantees: immutable; is valid as declared in {@link #isValidName(String)}
  */
-public class ProjectId {
-
-    public static final String MESSAGE_CONSTRAINTS =
-            "Project names should only contain alphanumeric characters and spaces, and it should not be blank";
-
-    /*
-     * The first character of the address must not be a whitespace,
-     * otherwise " " (a blank string) becomes a valid input.
-     */
-    public static final String VALIDATION_REGEX = "[\\p{Alnum}][\\p{Alnum} ]*";
-
-    public final String fullId;
+public class ProjectId extends Id {
 
     /**
      * Constructs a {@code Id}.
@@ -26,42 +14,7 @@ public class ProjectId {
      * @param id A valid Id.
      */
     public ProjectId(String id) {
-        requireNonNull(id);
-        checkArgument(isValidId(id), MESSAGE_CONSTRAINTS);
-        fullId = id;
-    }
-
-    /**
-     * Returns true if a given string is a valid name.
-     */
-    public static boolean isValidId(String test) {
-        return test.matches(VALIDATION_REGEX);
-    }
-
-
-    @Override
-    public String toString() {
-        return fullId;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof ProjectId)) {
-            return false;
-        }
-
-        ProjectId otherProjectId = (ProjectId) other;
-        return fullId.equals(otherProjectId.fullId);
-    }
-
-    @Override
-    public int hashCode() {
-        return fullId.hashCode();
+        super(id);
     }
 
 }
